Adapt solution heavily inspired from wenzhengjiang
With this username there is a different repository having solutions
already done, I could not figure out what I was doing wrong at the end,
my solution was a complete mess. Mainly wrong that I was not checking
for the adress being outside of stack range and not handling copyout
correctly.