Replease ref decrement with kfree
Since we don't hold the lock at the point of decrementing the ref, it's
safer to call kfree which, in the case that the process had been
scheduled out in the meantime and the references have changed, can also
free the page. If the scheduler does no funny things this should not
happen since if kget_ref == 1 another branch of the code is run
through, but usertests would give me missing free pages so this was the
only explanation (I could not track this down with gdb).